backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
tortoise-orm==0.24.2
aiogram==3.18.0
asyncpg==0.30.0
python-dotenv==1.0.1
